+++ f64/doc.go
@@ -1,0 +1,2 @@
+// Package f64 provides validators for types.Float64 attributes.
+package f64
