--- float64/doc.go
@@ -1,2 +1,0 @@
-// Package float64 provides validators for types.Float64 attributes.
-package float64
