+++ i64/doc.go
@@ -1,0 +1,2 @@
+// Package i64 provides validators for types.Int64 attributes.
+package i64
