--- int64/doc.go
@@ -1,2 +1,0 @@
-// Package int64 provides validators for types.Int64 attributes.
-package int64
