+++ str/doc.go
@@ -1,0 +1,2 @@
+// Package str provides validators for types.String attributes.
+package str
