--- string/doc.go
@@ -1,2 +1,0 @@
-// Package string provides validators for types.String attributes.
-package string
